Updated match-users to deal with both groups and individual users, tests updated to check all keys given, need to add tests for groups as well
