+++ Models/ExtendedSchemas.py
@@ -1,0 +1,7 @@
+from marshmallow import Schema, fields, ValidationError
+
+class MatchUsersSchema(Schema):
+    UserId = fields.Integer(required=True)
+    TripStartLocation = fields.List(fields.String(), required=True)
+    TripStopLocation = fields.List(fields.String(), required=True)
+    time = fields.Float(required=False)
